Added logic and assets for dyanamic background
